src/main/java/com/fleetScan/FleetscanAiDetection/fleet_ai/requirements.txt
fastapi
uvicorn[standard]
easyocr
torch
torchvision
python-multipart
pillow
numpy
opencv-python
ultralytics
